Изменен тип sx->tc на size_t

+++ libs/compiler/analyzer.h
@@ -80,7 +80,7 @@ typedef struct analyzer
 
 	int buf_flag;
 	int buf_cur;
-	int temp_tc;
+	size_t temp_tc;
 	int error_flag;
 } analyzer;
